moved activation to UserJwtResource and whereby adjusted tests settings
