+++ src/test/java/com/devcomanda/easymeetup/controller/ActivationUserTest.java
@@ -1,13 +1,17 @@
 package com.devcomanda.easymeetup.controller;
 
+import com.devcomanda.easymeetup.controller.security.UserJwtResource;
 import com.devcomanda.easymeetup.model.security.InvalidUserActivationKeyException;
 import com.devcomanda.easymeetup.service.UserSecurityService;
 import com.devcomanda.easymeetup.service.UserService;
+import com.devcomanda.easymeetup.service.security.jwt.TokenProvider;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
@@ -22,7 +26,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
  */
 
 @RunWith(SpringRunner.class)
-@WebMvcTest(value = UserController.class, secure = false)
+@WebMvcTest(value = UserJwtResource.class, secure = false)
 public class ActivationUserTest {
 
     @Autowired
@@ -33,6 +37,13 @@ public class ActivationUserTest {
 
     @MockBean
     private UserService userService;
+
+    @MockBean
+    private TokenProvider tokenProvider;
+
+    @MockBean
+    @Qualifier("apiAuthManager")
+    private AuthenticationManager authenticationManager;
 
     @Test
     public void shouldActivateNewUser() throws Exception{
